+++ opentreemap/treemap/templates/treemap/map-browse-trees.html
@@ -40,7 +40,7 @@
       <div class="accordion-inner benefit-values hidden">
         {% comment %}
         TODO: on tree select, make ajax call to ecobenefits with
-         just one tree. render, calling _format_benefits with
+         just one tree. render, calling format_benefits with
          summary = True, and stick the results from
          search_tree_benefits in here.
         {% endcomment %}
