+++ opentreemap/treemap/urls.py
@@ -13,7 +13,7 @@ from treemap.views import (boundary_to_geojson_view, index_view, map_view,
                            get_map_feature_detail_view, edits_view,
                            search_tree_benefits_view, species_list_view,
                            boundary_autocomplete_view, instance_user_view,
-                           map_feature_popup_view, instance_user_audits,
+                           map_feature_popup_view, instance_user_audits_view,
                            plot_accordion_view, add_map_feature_view,
                            add_tree_photo_endpoint, photo_review_endpoint,
                            photo_review_partial_endpoint,
@@ -89,5 +89,5 @@ urlpatterns = patterns(
     url(r'^benefit/search$', search_tree_benefits_view),
     url(r'^users/%s/$' % USERNAME_PATTERN, instance_user_view,
         name="user_profile"),
-    url(r'^users/%s/edits/$' % USERNAME_PATTERN, instance_user_audits),
+    url(r'^users/%s/edits/$' % USERNAME_PATTERN, instance_user_audits_view),
 )
